+++ deep8/Deep8.h
@@ -1,23 +1,23 @@
 #ifndef DEEP8_DEEP8_H
 #define DEEP8_DEEP8_H
 
-#include "basic/Basic.h"
-#include "basic/Exception.h"
-#include "basic/CudaException.h"
-#include "basic/MemoryAllocator.h"
-#include "basic/Shape.h"
-
-#include "utils/Utils.h"
-#include "utils/MathUtils.h"
-#include "utils/CudaMathUtils.h"
-
-#include "basic/MemoryPool.h"
-#include "basic/Device.h"
-#include "basic/TensorStorage.h"
-#include "basic/Tensor.h"
-
-#include "utils/ShapeUtils.h"
-#include "utils/TensorUtils.h"
+//#include "basic/Basic.h"
+//#include "basic/Exception.h"
+//#include "basic/CudaException.h"
+//#include "basic/MemoryAllocator.h"
+//#include "basic/Shape.h"
+//
+//#include "utils/Utils.h"
+//#include "utils/MathUtils.h"
+//#include "utils/CudaMathUtils.h"
+//
+//#include "basic/MemoryPool.h"
+//#include "basic/Device.h"
+//#include "basic/TensorStorage.h"
+//#include "basic/Tensor.h"
+//
+//#include "utils/ShapeUtils.h"
+//#include "utils/TensorUtils.h"
 
 #include "nodes/Node.h"
 #include "nodes/Variable.h"
