+++ deep8/basic/TensorStorage.h
@@ -1,6 +1,7 @@
 #ifndef DEEP8_TENSORSTORAGE_H
 #define DEEP8_TENSORSTORAGE_H
 
+#include "Device.h"
 
 namespace Deep8 {
 
@@ -23,36 +24,11 @@ public:
     /**the memory size*/
     size_t size;
 
-    explicit TensorStorage(): ptr(nullptr), refPtr(nullptr), size(0), device(nullptr) {
-    }
+	explicit TensorStorage();
+	explicit TensorStorage(void *p, size_t *refP, size_t s, Device *d);
+	explicit TensorStorage(const TensorStorage &other);
 
-    explicit TensorStorage(void *p, size_t *refP, size_t s, Device *d)
-            : ptr(p), refPtr(refP), size(s), device(d) {
-        DEEP8_ARGUMENT_CHECK(nullptr != p && nullptr != refP && s > 0 && nullptr != device, "the memory is error");
-
-        (*refPtr) = 1;
-    }
-
-	explicit TensorStorage(const TensorStorage &other) {
-		if (nullptr != other.refPtr) {
-			(*other.refPtr)++;
-		}
-		
-		ptr    = other.ptr;
-		refPtr = other.refPtr;
-		size   = other.size;
-		device = other.device;
-	}
-
-    ~TensorStorage() {
-        if (nullptr != refPtr) {
-            (*refPtr)--;
-
-            if (0 == (*refPtr)) {
-                free();
-            }
-        }
-    }
+	~TensorStorage();
 
     TensorStorage &operator=(const TensorStorage &other) {
 		if (nullptr != other.refPtr) {
@@ -77,31 +53,9 @@ public:
 
 protected:
 
-    void free() {
-        if (DeviceType::CPU == device->type) {
-            device->free(ptr);
-            device->free(refPtr);
-
-            ptr    = nullptr;
-            refPtr = nullptr;
-            size   = 0;
-        } else {
-#ifdef HAVE_CUDA
-            auto gpuDevice = static_cast<GPUDevice*>(device);
-
-            gpuDevice->free(ptr);
-            gpuDevice->freeCPU(refPtr);
-
-            ptr    = nullptr;
-            refPtr = nullptr;
-            size   = 0;
-#else
-            DEEP8_RUNTIME_ERROR("can not call GPU function withou a GPU");
-#endif
-        }
-
-        device = nullptr;
-    }
+	void freeCPU();
+	void freeGPU();
+	void free();
 };
 
 }
