+++ deep8/nodes/Function.cpp
@@ -3,6 +3,34 @@
 namespace Deep8 {
 
 template <typename T>
+void Function<T>::forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) {
+	DEEP8_RUNTIME_ERROR("can not call this forwardCPU by Function class");
+}
+
+template <typename T>
+void Function<T>::forwardGPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) {
+	DEEP8_RUNTIME_ERROR("can not call this forwardGPU by Function class");
+}
+
+template <typename T>
+void Function<T>::backwardCPU(const std::vector<const Tensor<T>*> &inputs,
+	const Tensor<T> *output,
+	const Tensor<T> *outputGradient,
+	size_t index,
+	Tensor<T> *iGradient) {
+	DEEP8_RUNTIME_ERROR("can not call this backwardCPU by Function class");
+}
+
+template <typename T>
+void Function<T>::backwardGPU(const std::vector<const Tensor<T>*> &inputs,
+	const Tensor<T> *output,
+	const Tensor<T> *outputGradient,
+	size_t index,
+	Tensor<T> *iGradient) {
+	DEEP8_RUNTIME_ERROR("can not call this backwardGPU by Function class");
+}
+
+template <typename T>
 void Function<T>::forward() {
 	DEEP8_ARGUMENT_CHECK(NodeType::Variable == this->output->type, "the output must be a Variable type");
