+++ deep8/nodes/Function.h
@@ -51,29 +51,10 @@ protected:
     }
 
 protected:
-	virtual void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) {
-		DEEP8_RUNTIME_ERROR("can not call this forwardCPU by Function class");
-	}
-
-	virtual void forwardGPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) {
-		DEEP8_RUNTIME_ERROR("can not call this forwardGPU by Function class");
-	}
-
-	virtual void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-		const Tensor<T> *output,
-		const Tensor<T> *outputGradient,
-		size_t index,
-		Tensor<T> *iGradient) {
-		DEEP8_RUNTIME_ERROR("can not call this backwardCPU by Function class");
-	}
-
-	virtual void backwardGPU(const std::vector<const Tensor<T>*> &inputs,
-		const Tensor<T> *output,
-		const Tensor<T> *outputGradient,
-		size_t index,
-		Tensor<T> *iGradient) {
-		DEEP8_RUNTIME_ERROR("can not call this backwardGPU by Function class");
-	}
+	virtual void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output);
+	virtual void forwardGPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output);
+	virtual void backwardCPU(const std::vector<const Tensor<T>*> &inputs, const Tensor<T> *output, const Tensor<T> *outputGradient, size_t index, Tensor<T> *iGradient);
+	virtual void backwardGPU(const std::vector<const Tensor<T>*> &inputs, const Tensor<T> *output, const Tensor<T> *outputGradient, size_t index, Tensor<T> *iGradient);
 
 public:
 	void forward() override;
