+++ deep8/nodes/InputParameter.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_INPUTPARAMETER_H
 #define DEEP8_INPUTPARAMETER_H
 
+#include "Parameter.h"
+
 namespace Deep8 {
 
 /**
@@ -20,26 +22,11 @@ public:
      * feed the data into the InputParameter Node
      * the pointer's memory must bigger than the value size
      */
-    void feed(const T *ptr) {
-        DEEP8_ARGUMENT_CHECK(nullptr != ptr, "the pointer can not be null");
+	void feed(const T *ptr);
 
-        if (this->value.device()->type == DeviceType::CPU) {
-            this->value.device()->copy(ptr, this->value.raw(), sizeof(T) * this->value.size());
-        } else {
-#ifdef HAVE_CUDA
-            static_cast<GPUDevice*>(value.device())->copyFromCPUToGPU(ptr, value.raw(), sizeof(T) * value.size());
-#else
-            DEEP8_RUNTIME_ERROR("can not call a GPU function without a GPU");
-#endif
-        }
-    }
+	void zeroGradient() override;
 
-    void zeroGradient() override {
-    }
-
-	bool isScalar() override {
-		return this->value.isScalar();
-	}
+	bool isScalar() override;
 };
 
 }
